+++ src/ria/hid/pad.h
@@ -22,17 +22,17 @@ void pad_stop(void);
 bool pad_xreg(uint16_t word);
 
 // Parse HID report descriptor for gamepad.
-bool pad_mount(uint8_t slot, uint8_t const *desc_data, uint16_t desc_len,
+bool pad_mount(int slot, uint8_t const *desc_data, uint16_t desc_len,
                uint16_t vendor_id, uint16_t product_id);
 
 // Clean up descriptor when device is disconnected.
-bool pad_umount(uint8_t slot);
+bool pad_umount(int slot);
 
 // Process HID gamepad report.
-void pad_report(uint8_t slot, uint8_t const *data, uint16_t len);
+void pad_report(int slot, uint8_t const *data, uint16_t len);
 
-void pad_home_button(uint8_t slot, bool pressed);
+void pad_home_button(int slot, bool pressed);
 
-int pad_get_player_num(uint8_t slot);
+int pad_get_player_num(int slot);
 
 #endif /* _PAD_H_ */
